распределила модули и точку входа
