+++ js/utils/get-two-digits-from-01-to-10.js
@@ -1,0 +1,8 @@
+import {getRandomPositiveInteger} from './get-random-positive-integer.js';
+
+const getTwoDigitsFrom01To10 = () => {
+  const number = getRandomPositiveInteger(1, 10);
+  return number < 10 ? `0${number}` : `${number}`;
+};
+
+export {getTwoDigitsFrom01To10};
